Adding redux-thunk middleware to fetch categories.
