+++ package.json
@@ -35,6 +35,7 @@
     "react-toastify": "^9.1.1",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.4.2",
     "reselect": "^4.1.7",
     "sass": "^1.58.1",
     "sequelize": "^6.28.0"
